+++ py_clickhouse_migrator/cli/__init__.py
@@ -1,0 +1,1 @@
+from py_clickhouse_migrator.cli.app import main as main